64bit compatibility
git-svn-id: svn://localhost/trunk@985 3bdad9f7-c964-442a-868d-553b97b55ca0

+++ surfit/src/sstuff/ptypes/pport.h
@@ -72,7 +72,7 @@ PTYPES_BEGIN
 // useful typedefs
 //
 
-typedef unsigned int    uint;
+typedef size_t    uint;
 typedef unsigned long   ulong;
 typedef unsigned short  ushort;
 typedef unsigned char   uchar;
